[catalog] moved the location catalog logic to the location dao.
