+++ cwms_radar_api/src/main/java/cwms/radar/data/CwmsDataManager.java
@@ -304,106 +304,7 @@ public class CwmsDataManager implements AutoCloseable {
         return cat;
     }
 
-    public Catalog getLocationCatalog(String cursor, int pageSize, String unitSystem, Optional<String> office) {
-        int total = 0;
-        String locCursor = "*";
-        if( cursor == null || cursor.isEmpty() ){
-            SelectJoinStep<Record1<Integer>> count = dsl.select(count(asterisk())).from(AV_LOC);
-            if( office.isPresent() ){
-                count.where(AV_LOC.DB_OFFICE_ID.upper().eq(office.get().toUpperCase()));
-            }
-            total = count.fetchOne().value1().intValue();
-        } else {
-            logger.info("getting non-default page");
-            // get totally from page
-            String[] parts = Catalog.decodeCursor(cursor, "|||");
 
-            logger.info("decoded cursor: " + String.join("|||", parts));
-            for( String p: parts){
-                logger.info(p);
-            }
-
-            if(parts.length > 1) {
-                locCursor = parts[0].split("\\/")[1];
-                total = Integer.parseInt(parts[1]);
-            }
-        }
-
-        SelectConditionStep<Record1<String>> tmp = dsl.select(AV_LOC.LOCATION_ID)
-                               .from(AV_LOC)
-                               .where(AV_LOC.LOCATION_ID.greaterThan(locCursor))
-                               .and(AV_LOC.UNIT_SYSTEM.eq(unitSystem));
-        if( office.isPresent()){
-            tmp = tmp.and(AV_LOC.DB_OFFICE_ID.upper().eq(office.get().toUpperCase()));
-        }
-        Table<?> forLimit = tmp.orderBy(AV_LOC.BASE_LOCATION_ID).limit(pageSize).asTable();
-        SelectConditionStep<Record> query = dsl.select(
-                                    AV_LOC.asterisk(),
-                                    AV_LOC_GRP_ASSGN.asterisk()
-                                )
-                                .from(AV_LOC)
-                                .innerJoin(forLimit).on(forLimit.field(AV_LOC.LOCATION_ID).eq(AV_LOC.LOCATION_ID))
-                                .leftJoin(AV_LOC_GRP_ASSGN).on(AV_LOC_GRP_ASSGN.LOCATION_ID.eq(AV_LOC.LOCATION_ID))
-                                .where(AV_LOC.UNIT_SYSTEM.eq(unitSystem))
-                                .and(AV_LOC.LOCATION_ID.upper().greaterThan(locCursor));
-
-        if( office.isPresent() ){
-            query.and(AV_LOC.DB_OFFICE_ID.upper().eq(office.get().toUpperCase()));
-        }
-        query.orderBy(AV_LOC.LOCATION_ID);
-        logger.info( query.getSQL(ParamType.INLINED));
-        HashMap<usace.cwms.db.jooq.codegen.tables.records.AV_LOC, ArrayList<usace.cwms.db.jooq.codegen.tables.records.AV_LOC_ALIAS>> theMap = new HashMap<>();
-        //Result<?> result = query.fetch();
-        query.fetch().forEach( row -> {
-            usace.cwms.db.jooq.codegen.tables.records.AV_LOC loc = row.into(AV_LOC);
-            if( !theMap.containsKey(loc)){
-                theMap.put(loc, new ArrayList<>() );
-            }
-            usace.cwms.db.jooq.codegen.tables.records.AV_LOC_ALIAS alias = row.into(AV_LOC_ALIAS);
-            usace.cwms.db.jooq.codegen.tables.records.AV_LOC_GRP_ASSGN group = row.into(AV_LOC_GRP_ASSGN);
-            if( group.getALIAS_ID() != null ){
-                theMap.get(loc).add(alias);
-            }
-        });
-
-        List<? extends CatalogEntry> entries =
-        theMap.entrySet().stream().map( e -> {
-            logger.info(e.getKey().toString());
-            LocationCatalogEntry ce = new LocationCatalogEntry(
-                e.getKey().getDB_OFFICE_ID(),
-                e.getKey().getLOCATION_ID(),
-                e.getKey().getNEAREST_CITY(),
-                e.getKey().getPUBLIC_NAME(),
-                e.getKey().getLONG_NAME(),
-                e.getKey().getDESCRIPTION(),
-                e.getKey().getLOCATION_KIND_ID(),
-                e.getKey().getLOCATION_TYPE(),
-                e.getKey().getTIME_ZONE_NAME(),
-                e.getKey().getLATITUDE() != null ? e.getKey().getLATITUDE().doubleValue() : null,
-                e.getKey().getLONGITUDE() != null ? e.getKey().getLONGITUDE().doubleValue() : null,
-                e.getKey().getPUBLISHED_LATITUDE() != null ? e.getKey().getPUBLISHED_LATITUDE().doubleValue() : null,
-                e.getKey().getPUBLISHED_LONGITUDE() != null ? e.getKey().getPUBLISHED_LONGITUDE().doubleValue() : null,
-                e.getKey().getHORIZONTAL_DATUM(),
-                e.getKey().getELEVATION(),
-                e.getKey().getUNIT_ID(),
-                e.getKey().getVERTICAL_DATUM(),
-                e.getKey().getNATION_ID(),
-                e.getKey().getSTATE_INITIAL(),
-                e.getKey().getCOUNTY_NAME(),
-                e.getKey().getBOUNDING_OFFICE_ID(),
-                e.getKey().getMAP_LABEL(),
-                e.getKey().getACTIVE_FLAG().equalsIgnoreCase("T") ? true : false,
-                e.getValue().stream().map( a -> {
-                    return new LocationAlias(a.getCATEGORY_ID()+"-"+a.getGROUP_ID(),a.getALIAS_ID());
-                }).collect(Collectors.toList())
-            );
-
-            return ce;
-        }).collect(Collectors.toList());
-
-        Catalog cat = new Catalog(cursor,total,pageSize,entries);
-        return cat;
-    }
 
 
 }
